Add files via upload
This notebook shows an example of how to configure and use Trusted Context in the Data Virtualization database. After some initial set up cells, the first part of of the notebook is the configuraiton section that would need to be run once (Configure Trusted Context), and the second part shows how to use it and switch users from the one that made the connection (Example of Using Trusted Context).